fix: update eda notebook to display report
In the original notebook, when it was pushed to GitHub the pandas-profiling
report does not display. Here I updated the code to use the `to_notebook_iframe()`
method which converts it into HTML format inside the notebook.

Note that I also had to downgrade the pandas-profiling package to version
3.0.0 as 3.1.0 had some issues with displaying the report properly.

I also fixed the link to the original dataset in the notebook.